Remove EDH posts api methods & tests
